Wandb, train-val loops, new gta5 dataset

+++ datasets/cityscapes.py
@@ -1,4 +1,5 @@
 from torch.utils.data import Dataset
+import numpy as np
 
 import os
 import cv2
@@ -26,10 +27,10 @@ class CityScapes(Dataset):
             for img_path in os.listdir(img_city_dir): # frankfurt_000000_000294_leftImg8bit.png
                 if img_path.endswith(".png"):
                   self.imgs_path.append(os.path.join(img_city_dir, img_path)) # ./images/train/frankfurt/frankfurt_000000_000294_leftImg8bit.png
-                  
+
                   target_color_path = img_path.replace("leftImg8bit", "gtFine_color") # frankfurt_000000_000294_gtFine_color.png
                   target_labelIds_path = img_path.replace("leftImg8bit", "gtFine_labelTrainIds") # frankfurt_000000_000294_gtFine_labelTrainIds.png
-                  
+
                   self.targets_color_path.append(os.path.join(target_city_dir, target_color_path)) # ./gtFine/train/frankfurt/frankfurt_000000_000294_gtFine_color.png
                   self.targets_labelIds_path.append(os.path.join(target_city_dir, target_labelIds_path)) # ./gtFine/train/frankfurt/frankfurt_000000_000294_gtFine_labelTrainIds.png
 
@@ -37,7 +38,7 @@ class CityScapes(Dataset):
         image = Image.open(self.imgs_path[idx]).convert('RGB')
 
         target_color = Image.open(self.targets_color_path[idx]).convert('RGB')
-        target_labelIds = cv2.imread(self.targets_labelIds_path[idx], cv2.IMREAD_UNCHANGED)
+        target_labelIds = cv2.imread(self.targets_labelIds_path[idx], cv2.IMREAD_UNCHANGED).astype(np.long)
 
         if self.transform is not None:
             image = self.transform(image)
@@ -46,6 +47,6 @@ class CityScapes(Dataset):
             target_labelIds = self.target_transform(target_labelIds)
 
         return image, target_color, target_labelIds
-        
+
     def __len__(self):
         return len(self.imgs_path)
